requirements.txt
Flask
mysql-connector-python
pytesseract
fuzzywuzzy
Pillow